requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.20.0
google-ads>=25.0.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.124.0
schedule==1.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
requests==2.31.0
pytz==2024.1